Fix potentially-contentious sometimes-incoherence
"syntax", in programming, is generally a superset of "grammar",
which is a very specific thing (usually defined in BNF or what).

"syntax", in linguistics, is a subset of "grammar".

So depending on which viewpoint one has, the phrase either makes
sense or is incoherent. Aaaargh. So I rephrased it.

Thanks Léa!