Fix presentation of Zhaed rel. common tongues.
Relevant excerpts from discussion with mako:

14:00:39 <mako yass>
1: What are the common tongues, are they natural language.
2: Is ❝prose❞ a special term.
3: what exactly is the nature of magic here? I'm not sure whether your
handling of it here makes sense, I just can't possibly tell. This is not
fun for rationalists.

14:01:21 <mako yass>
Bah, OK, most forms of confusion are fun for rationalists.

14:01:57 <passcod>
common tongues are natural, yes. there's five

14:02:16 <mako yass>
But there need to be consistent rules available for application, or
inference is not possible.

14:02:37 <passcod> haven't decided names for them yet, but I do know
where they come from, and that's not relevant yet

14:04:08 <passcod> 2: no. "a non-poem-ish-looking block of gibberish
text" would have been too long

14:06:03 <mako yass>
OK it's kind of hard to believe the general public would mistake majikk
for straight writ. That's the kind of thing that can be dangerous, ergo,
they'd be vigilant about it. Though maybe it's not hard to believe an out
of touch professor would have a misconception about a misconception.
It's a stretch.

14:07:33 <mako yass>
All contingent on your answer to 3, really.

14:11:34 <mako yass>
I'm really curious as to how using Zhaed symbols to write a spoken
language could fail to work. I'm imagining 6 year olds accidentally
writing Zhaed symbols during handwriting class and being mysteriously
unable to cogniss their mistake.

14:12:51 <mako yass>
Hah. So like, particular errors make the results disproportionately harder
to read than others.

14:17:35 <passcod>
Hmm. My original idea was to have a ~10M BCE point in time when there
wasn't magic, then have a Ra-like advance, essentially making WA sci-fi.
That can be taken out in favour of an always-magic universe, which would
align with HPJEV's Atlantis hypothesis: spells and such are triggers for
the underlying VM, which runs a bit closer to the bare metal, and people
have forgotten (or the knowledge has been purged intentionally) how to
access the layer(s) beneath it. After that, you've got several millenia
of more or less advanced civilisations rising and collapsing and creating
more and more layers and versions of using the original Atlantis VM.
(Not sure if my VM is named Atlantis or not. I might just.)

14:18:13 <passcod>
s/, which/. And it/

14:18:57 <passcod>
s/several millenia/a loooong time/

14:21:11 <passcod>
So, yes, that means I can do anything. But there's (been) a lot of
knowledge loss, so the current situation is akin to researching how
our universe works. But with magic. Which really doesn't change much.

14:23:35 <passcod>
I mean, people don't use magic freely everyday. They use bottled spells
and programmed tech and enchanted devices that are manufactured,
designed, crafted, etc by people who actually know how to do that.

14:25:51 <passcod>
It's like computers nowadays. People don't blast machine code from
their minds straight into a box to fetch a tweet. There's hundreds of
layers of abstraction that make it as easy as clicking a button, and
clicking that button won't make your computer explode or the US nuclear
war heads to detonate.

14:35:21 <passcod>
And the Zhaed symbol-set was inspired from something else. Anyway,
activating Zhaed requires knowledge of how to do so and an energy source,
which is something 6yos won't have. The danger isn't there. The danger
is like if you have a variable called `onc` and you mistype it as `once`
because your brain latches on to words it knows. But `once` could be
just a compiler error, or it could be the name of a routine you didn't
want to run, and you end up feeding `/` to `rm -rf` or something.

14:37:50 <mako yass>
if onc meant something significantly different from once... it could be
that the difference is subtle, linguistic forces would have dragged
modern languages closer to our intuitions, so yeah, the old meaning
could be hard to grasp.

14:39:23 <passcod>
so? not too disappointed with my "I can do anything" (not really) approach
to fictional magic?

14:40:15 <mako yass>
I'm pretty sure there had to be a VM+energy source somewhere.

[...]

14:50:41 <passcod>
Now that your questions are answered, what's wrong with that paragraph? ;)

14:52:46 <mako yass>
I can't imagine a pathway by which a spoken language could emerge from
a written magic language, in which many of the words and approximate
grammar are retained.

14:52:56 <mako yass>
without retaining a lot of the meanings.

14:57:27 <passcod>
oh, it wouldn't. the spoken language was already there long before
Zhaed came along. There was even small-audience, fragmented writ
versions. Some used scripts similar to Zhaed, or maybe even directly
inspred from it. Zhaed was merely used as the alphabet for the language
when a more unified authority came along and sought to standardise it all.
Zhaed itself isn't pronounced. There /are/ names in the common tongues
to describe some of the symbols, but they're not proper to Zhaed. It's
like 'boat' and 'bateau' and '船'.

15:00:09 <passcod>
The alphabet is pronounced, yes, but the aphabet isn't Zhaed. It merely
resembles it so closely it can be dangerous. Zhaed is a bit as if the
Old English Alphabet was a symbol-set for magic. Or like nordic runes,
which have names in finnish and swedish and other languages, but are
(probably, not quite sure) pronounced.

15:00:28 <passcod>
*not

15:01:38 <passcod>
In fact, I think if the original designers of Zhaed lived to see it used
as a basis for a spoken language, they would be livid! :D

15:03:21 <mako yass>
def needs a correction here then

- the resemblance occurs because the dominant language from which ours
are derived was itself derived from the Zhaed symbol set

+ the resemblance occurs because the alphabet of the dominant language
from which ours are derived was itself derived from the Zhaed symbol set

The bad interpretation is a category error but that could just lead
the reader to be confused about the categories.

15:03:41 <mako yass>
You should probably name their language there.

15:04:31 <mako yass>
>the resemblance occurs because the alphabet of the <old indo-germanic>
was itself derived from the Zhaed symbol set

15:04:47 <mako yass>
[assume they can follow their linguistic lineage]

15:05:02 <mako yass>
[especially given the context of the assertion]

15:05:35 <passcod>
ah, yes

15:05:42 <passcod>
thanks

[...]

15:11:09 <mako yass>
Also when you say danger I, as a new reader, thought you meant actual
danger of catching things on fire and summoning demons rather than
say, danger of being a bit more confused than otherwise.